requirements.txt
pyqt==5.15.6
numpy==1.21.2
scipy==1.8.0
matplotlib==3.5.1
ipywidgets==7.6.5

